add clear all completed tasks function to test this feature
